Refattorei input
